// src/types.ts
// Generated by https://quicktype.io

export interface User {
    repositories: Repositories;
}

export interface Repositories {
    nodes: NodeElement[];
}

export interface NodeElement {
    name:      string;
    languages: Languages;
    visibility: string;
    isTemplate: boolean;
    [key: string]: any;
}

export interface Languages {
    edges: Edge[];
}

export interface Edge {
    size: number;
    node: EdgeNode;
}

export interface EdgeNode {
    name: string;
    color: string;
}


export interface TreeData {
    node: Node;
}

export interface Node {
    entries: Entry[];
}

export interface Entry {
    language: Language | null;
    lineCount:     number;
    object: {
        id: string;
    }
}

export interface Language {
    name: string;
}